+++ nutype/src/lib.rs
@@ -177,7 +177,7 @@
 //! ### String derivable traits
 //!
 //! The following traits can be derived for a string-based type:
-//! `Debug`, `Clone`, `PartialEq`, `Eq`, `PartialOrd`, `Ord`, `FromStr`, `AsRef`, `From`, `TryFrom`, `Into`, `Hash`, `Borrow`, `Display`, `Serialize`, `Deserialize`.
+//! `Debug`, `Clone`, `PartialEq`, `Eq`, `PartialOrd`, `Ord`, `FromStr`, `AsRef`, `From`, `TryFrom`, `Into`, `Hash`, `Borrow`, `Display`, `Default`, `Serialize`, `Deserialize`.
 //!
 //!
 //! ## Integer
@@ -201,7 +201,7 @@
 //! ### Integer derivable traits
 //!
 //! The following traits can be derived for an integer-based type:
-//! `Debug`, `Clone`, `Copy`, `PartialEq`, `Eq`, `PartialOrd`, `Ord`, `FromStr`, `AsRef`, `Into`, `From`, `TryFrom`, `Hash`, `Borrow`, `Display`, `Serialize`, `Deserialize`.
+//! `Debug`, `Clone`, `Copy`, `PartialEq`, `Eq`, `PartialOrd`, `Ord`, `FromStr`, `AsRef`, `Into`, `From`, `TryFrom`, `Hash`, `Borrow`, `Display`, `Default`, `Serialize`, `Deserialize`.
 //!
 //!
 //! ## Float
@@ -226,7 +226,7 @@
 //! ### Float derivable traits
 //!
 //! The following traits can be derived for a float-based type:
-//! `Debug`, `Clone`, `Copy`, `PartialEq`, `PartialOrd`, `FromStr`, `AsRef`, `Into`, `From`, `TryFrom`, `Hash`, `Borrow`, `Display`, `Serialize`, `Deserialize`.
+//! `Debug`, `Clone`, `Copy`, `PartialEq`, `Eq`, `PartialOrd`, `Ord`, `FromStr`, `AsRef`, `Into`, `From`, `TryFrom`, `Hash`, `Borrow`, `Display`, `Default`, `Serialize`, `Deserialize`.
 //!
 //! It's also possible to derive `Eq` and `Ord` if the validation rules guarantee that `NaN` is excluded.
 //! This can be done by applying `finite` validation. For example:
@@ -285,6 +285,30 @@
 //!     // A fancy way to verify if the first character is uppercase
 //!     name.chars().next().map(char::is_uppercase).unwrap_or(false)
 //! }
+//! ```
+//! ## Deriving recipes
+//!
+//! ### Deriving `Default`
+//!
+//! ```
+//! use nutype::nutype;
+//!
+//! #[nutype(default = "Anonymous")]
+//! #[derive(Default)]
+//! pub struct Name(String);
+//! ```
+//!
+//! ### Deriving `Eq` and `Ord` on float types
+//!
+//! With nutype it's possible to derive `Eq` and `Ord` if there is `finite` validation set.
+//! The `finite` validation ensures that the valid value excludes `NaN`.
+//!
+//! ```
+//! use nutype::nutype;
+//!
+//! #[nutype(validate(finite))]
+//! #[derive(PartialEq, Eq, PartialOrd, Ord)]
+//! pub struct Weight(f64);
 //! ```
 //!
 //! ## How to break the constraints?