+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### v0.3.0 - 2023-??-??
+### v0.3.0 - 2023-06-25
 * [BREAKING] `min_len` and `max_len` validators run against number of characters in a string (`val.chars().count()`), not number of bytes (`val.len()`).
 * Add `finite` validation for float types which checks against NaN and infinity.
 * Support deriving of `Default`